feat: Make platform-specific crates optional
This changes the platform-specific packages to all be optional
dependencies, pulling in only the detected platform as required. This
reduces the amount of bytes people need to pull in regular usage. You
can still pull in alternative platform binaries by enabling the feature
for them, but this is probably not very useful.