fix: Name features according to Rust conventions instead of Go
This renames the features of `etcd-bin-vendored` for pulling in
platform-specific crates according to the Rust convention for the
`target_os`/`target_arch`. Previously, they were based on the Go
convention for this, which is what `etcd` binary releases are named
after. The crates themselves maintain their names because it seems
sensible to keep the wrapper package named after the binary it is
wrapping.